test(matrix-utils): split on_room_message tests into dedicated relay module
migrate on_room_message tests from tests/test_matrix_utils.py to
tests/test_matrix_utils_relay.py and split tests by domain

cleanup unused imports in core matrix_utils tests to reduce coupling
and improve maintainability

align fixtures and time mocks with the new test layout
